+++ src/router/index.js
@@ -24,22 +24,23 @@ const routes = [
     // which is lazy-loaded when the route is visited.
     component: () => import(/* webpackChunkName: "timeline" */ '../views/Timeline.vue')
   },
-  {
-    path: '/edge-connect',
-    name: 'edge-connect',
-    props: true,
-    // route level code-splitting
-    // this generates a separate chunk (about.[hash].js) for this route
-    // which is lazy-loaded when the route is visited.
-    component: () => import(/* webpackChunkName: "edge-connect" */ '../views/EdgeConnect.vue')
-  },
+  // {
+  // This might be connected again once we have better UX design
+  //   path: '/edge-connect',
+  //   name: 'edge-connect',
+  //   props: true,
+  //     route level code-splitting
+  //     this generates a separate chunk (about.[hash].js) for this route
+  //     which is lazy-loaded when the route is visited.
+  //   component: () => import(/* webpackChunkName: "edge-connect" */ '../views/EdgeConnect.vue')
+  // },
   {
     path: '/settings',
     name: 'settings',
     // route level code-splitting
     // this generates a separate chunk (about.[hash].js) for this route
     // which is lazy-loaded when the route is visited.
-    component: () => import(/* webpackChunkName: "edge-connect" */ '../views/EdgeConnect.vue') // '../views/Settings.vue')
+    component: () => import(/* webpackChunkName: "settings" */ '../views/Settings.vue')
   },
   {
     path: '/feedback',
@@ -64,14 +65,6 @@ const routes = [
     // this generates a separate chunk (about.[hash].js) for this route
     // which is lazy-loaded when the route is visited.
     component: () => import(/* webpackChunkName: "about" */ '../views/About.vue')
-  },
-  {
-    path: '/choose-edge-connection',
-    name: 'choose-edge-connection',
-    // route level code-splitting
-    // this generates a separate chunk (about.[hash].js) for this route
-    // which is lazy-loaded when the route is visited.
-    component: () => import(/* webpackChunkName: "choose-edge-connection" */ '../views/ChooseEdgeConnection.vue')
   }
 ]